Separate class into its own file
